+++ sources/packages/api/src-sql/Types/Add-offering-type-scholastic-standing.sql
@@ -1,6 +1,6 @@
 -- Add Scholastic Standing to sims.offering_types.
 ALTER TYPE sims.offering_types
 ADD
-    VALUE 'Scholastic Standing'
+    VALUE 'Scholastic standing'
 AFTER
     'Private';
