+++ sources/packages/api/src/database/entities/offering-status.type.ts
@@ -1,0 +1,8 @@
+/**
+ * Represents the status of an offering.
+ */
+export enum OfferingStatus {
+  Approved = "Approved",
+  Pending = "Pending",
+  Declined = "Declined",
+}
