+++ sources/packages/api/src/database/entities/offering.type.ts
@@ -5,7 +5,7 @@ export enum OfferingTypes {
   /**
    * Offering is available for all the students.
    */
-  public = "Public",
+  Public = "Public",
   /**
    * Offering was created to fulfill the need of
    * a particular student/application.
@@ -14,5 +14,5 @@ export enum OfferingTypes {
   /**
    * Offering created for a change in scholastic standing reported by institution.
    */
-  ScholasticStanding = "Scholastic Standing",
+  ScholasticStanding = "Scholastic standing",
 }
