+++ sources/packages/api/src/database/entities/program-status.type.ts
@@ -1,0 +1,9 @@
+/**
+ * Represents the approval status for a program also stored
+ * on column approval_status on table education_programs.
+ */
+export enum ProgramStatus {
+  Approved = "Approved",
+  Pending = "Pending",
+  Declined = "Declined",
+}
